+++ javax/src/main/java/me/panpf/javax/util/IndexedPredicate.java
@@ -19,5 +19,5 @@ package me.panpf.javax.util;
 import org.jetbrains.annotations.NotNull;
 
 public interface IndexedPredicate<T> {
-    boolean predicate(int index, @NotNull T t);
+    boolean accept(int index, @NotNull T t);
 }