+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.6.7
 
 * :hammer: The method in Stringx whose name ends with `or` adds a String overload method.
-* :sparkles: Stringx add reversed method
+* :sparkles: Stringx add reversed, filter, filterBlank method
 
 ## 0.6.6